+++ server/locale/GB/index.js
@@ -1,5 +1,5 @@
 import getMutations from './mutations';
-import logos from './logos';
+import logos from '../../message/logos';
 import styles from './styles';
 import validOptions from './validOptions';