+++ server/locale/US/GPL/mutations/gplq.js
@@ -1,4 +1,4 @@
-import Logo from '../logos';
+import Logo from '../../../../message/logos';
 import {
     textWrap,
     messageLogoWidth,
@@ -7,13 +7,13 @@ import {
     xSmallFallback,
     logo20x1
 } from '../../../../message/mediaQueries';
-import { flexLogoMutations, textLogoMutations } from './common';
+import { flexLogoMutations, textLogoMutations } from '../../../../message/logoMutations';
 
 const flex = [
     [
         'default',
         {
-            logo: Logo.PRIMARY.WHITE,
+            logo: Logo.PP_PAYPAL.WHITE,
             headline: [
                 {
                     tag: 'xsmall'
@@ -63,6 +63,7 @@ const flex = [
             ]
         }
     ],
+    ['color:white-no-border', { logo: Logo.PP_PAYPAL.COLOR }],
     ...flexLogoMutations
 ];
 
@@ -76,9 +77,10 @@ export default {
                     textWrap(textSize * 32, textSize, 'US'),
                     xSmallFallback(textSize * 16),
                     messageLogoWidth(false, textSize * 4, textSize * 1.25),
-                    setLogoTop(textSize * 16)
+                    setLogoTop(textSize * 16),
+                    `.message__headline > .tag--medium > span:not(.weak):first-child {white-space: nowrap;}`
                 ],
-                logo: Logo.PRIMARY.COLOR,
+                logo: Logo.PP_PAYPAL.COLOR,
                 headline: [
                     {
                         tag: 'medium',
@@ -94,7 +96,8 @@ export default {
             ({ textSize }) => ({
                 styles: [
                     `@media screen and (max-width: ${textSize *
-                        14.15}px) { .message__headline > .tag--medium > span > span:first-child { white-space: normal; } }`,
+                        18.5}px) { .message__headline > .tag--medium > span > span:first-child { white-space: normal; } }
+                    .message__headline > .tag--medium > span:not(.weak):first-child {white-space: nowrap;}`,
                     xSmallFallback(textSize * 10.75),
                     setLogoTop(textSize * 32),
                     messageLogoWidth(textSize * 6, textSize * 4, textSize * 1.25)
@@ -106,7 +109,8 @@ export default {
             ({ textSize }) => ({
                 styles: [
                     `@media screen and (max-width: ${textSize *
-                        14.15}px) { .message__headline > .tag--medium > span > span:first-child { white-space: normal; } }`,
+                        18.5}px) { .message__headline > .tag--medium > span > span:first-child { white-space: normal; } }
+                    .message__headline > .tag--medium > span:not(.weak):first-child {white-space: nowrap;}`,
                     xSmallFallback(textSize * 10.75),
                     messageLogoWidth(textSize * 6, textSize * 4, textSize * 1.25)
                 ]
@@ -122,7 +126,7 @@ export default {
                     altNoWrap(textSize * 10.6),
                     messageLogoWidth(textSize * 1.75, textSize * 4, textSize * 1.25)
                 ],
-                logo: Logo.PRIMARY.COLOR[0]
+                logo: Logo.PP_PAYPAL.COLOR[0]
             })
         ],
         [
@@ -147,7 +151,7 @@ export default {
             'logo.type:inline',
             ({ textSize }) => ({
                 styles: [xSmallFallback(textSize * 18), `.message__logo { width: ${textSize * 4}px }`],
-                logo: Logo.INLINE.COLOR,
+                logo: Logo.NO_PP_MONOGRAM.COLOR,
                 headline: [
                     {
                         tag: 'medium',