+++ src/utils/logger.js
@@ -16,13 +16,13 @@ export const logger = Logger({
     // Override transport so we can use withCredentials
     transport: ({ url, method, json, headers }) => {
         // Because there is no way to remove payload builders from beaver-logger
-        // Filter the meta object to remove inactive banner meta
+        // Filter the meta object to remove inactive banner meta commonly caused by SPAs
         const activeIndexes = json.events
             .map(({ payload: { index } }) => index)
             .concat(json.tracking.map(({ index }) => index));
 
         const trimmedMeta = objectKeys(json.meta)
-            .filter(index => activeIndexes.includes(index))
+            .filter(index => activeIndexes.includes(index.includes('modal') ? index.replace('modal-', '') : index))
             .reduce(
                 (accumulator, index) => ({
                     ...accumulator,
@@ -53,4 +53,11 @@ logger.addPayloadBuilder(payload => {
     delete payload.selector; // eslint-disable-line no-param-reassign
 
     return {};
+});
+
+logger.addTrackingBuilder(() => {
+    // Send a timestamp with every tracking event so they can be correctly ordered
+    return {
+        timestamp: new Date().getTime()
+    };
 });