+++ tests/functional/spec/US/accounts.js
@@ -11,5 +11,6 @@ export default [
     'DEV0000000PMG',
     'DEV0000000GPL',
     'DEV000000GPLQ',
-    'DEV00000GPLNQ'
+    'DEV00000GPLNQ',
+    'DEVGPLNQRANGE'
 ];