+++ tests/functional/spec/createBannerTest.js
@@ -3,15 +3,23 @@ import { configureToMatchImageSnapshot } from 'jest-image-snapshot';
 import { logScreenshot, logTestName } from './utils/logging';
 import selectors from './utils/selectors';
 
-const toMatchImageSnapshot = configureToMatchImageSnapshot({
+const toMatchTextSnapshot = configureToMatchImageSnapshot({
     failureThresholdType: 'percent',
-    failureThreshold: 0.002,
+    failureThreshold: 0.004,
     customDiffConfig: {
         threshold: 0.05
     }
 });
 
-expect.extend({ toMatchImageSnapshot });
+const toMatchFlexSnapshot = configureToMatchImageSnapshot({
+    failureThresholdType: 'percent',
+    failureThreshold: 0.005,
+    customDiffConfig: {
+        threshold: 0.05
+    }
+});
+
+expect.extend({ toMatchTextSnapshot, toMatchFlexSnapshot });
 
 const getConfigStrParts = (obj, keyPrefix = '') => {
     return Object.entries(obj).reduce((accumulator, [key, val]) => {
@@ -39,7 +47,7 @@ const getTestNameParts = (locale, { account, amount, style: { layout, ...style }
 };
 
 // returns height and width of banner in pixels
-const waitForBanner = async ({ testName, timeout }) => {
+const waitForBanner = async ({ testName, timeout, config }) => {
     try {
         const polling = 10;
         const result = await page.waitForFunction(
@@ -54,6 +62,14 @@ const waitForBanner = async ({ testName, timeout }) => {
                 if (iframe) {
                     const iframeBody = iframe.contentWindow.document.body;
                     const banner = iframeBody.querySelector(bannerSelectors.container);
+                    if (config?.style?.text?.align) {
+                        return (
+                            iframeBody?.clientHeight && {
+                                height: iframeBody.clientHeight,
+                                width: iframeBody.clientWidth
+                            }
+                        );
+                    }
                     return banner?.clientHeight && { height: banner.clientHeight, width: banner.clientWidth };
                 }
 
@@ -97,6 +113,7 @@ export default function createBannerTest(locale, testPage = 'banner.html') {
                     console.log(text);
                 }
             });
+            page.removeAllListeners('pageerror');
             page.on('pageerror', error => {
                 // TODO: find a way to re-launch the browser on error so tests can continue
                 // eslint-disable-next-line no-console
@@ -110,7 +127,7 @@ export default function createBannerTest(locale, testPage = 'banner.html') {
             await page.goto(`https://localhost.paypal.com:8080/snapshot/${testPage}?config=${JSON.stringify(config)}`);
             await waitForNavPromise;
 
-            const bannerDimensions = await waitForBanner({ testName, timeout: 2 * 1000 });
+            const bannerDimensions = await waitForBanner({ testName, timeout: 2 * 1000, config });
             expect(bannerDimensions.height).toBeGreaterThan(0);
             expect(bannerDimensions.width).toBeGreaterThan(0);
 
@@ -133,8 +150,9 @@ export default function createBannerTest(locale, testPage = 'banner.html') {
                 3
             );
 
+            const matchFunction = config?.style?.layout === 'text' ? 'toMatchTextSnapshot' : 'toMatchFlexSnapshot';
             const customSnapshotIdentifier = `${testNameParts.pop()}-${viewport.width}`;
-            expect(image).toMatchImageSnapshot({
+            expect(image)[matchFunction]({
                 diffDirection: snapshotDimensions.width > snapshotDimensions.height ? 'vertical' : 'horizontal',
                 customSnapshotsDir: ['./tests/functional/snapshots', ...testNameParts].join('/'),
                 customSnapshotIdentifier